refactor: 拆分 main.js，抽出 settingsRenderer、editableSchedule、scheduleGenerator 模組
main.js 從 1569 行縮減至 ~700 行，移除三大功能區塊：
- settingsRenderer.js：renderTasks/renderPersonnel/renderAll 等渲染函式
- editableSchedule.js：可編輯班表系統（拖拽、dropdown、save/cancel/exit）
- scheduleGenerator.js：renderFillStats/generateFullSchedule/displaySchedule

generateFullSchedule 改為接受 activeHolidays 參數，
各 call site 傳入 Array.from(activeHolidayDates)。
